Temporarily remove collaborator names ending in -
ASF Infra has a regex to test if a collaborator name is a valid Github username.  Github does not allow usernames to end in hyphen.  However, they used to do so and some usernames were grandfathered in.  However, ASF is not accepting those usernames.  I'll file a ticket with Infra but the "not valid username" is cluttering up the builds ML at the moment so I'm disabling until that gets resolved.